Delete VoiceChannelRouter, VoiceChannelV2 and V2 push support
