+++ Cartfile.resolved
@@ -5,7 +5,7 @@ github "wireapp/ocmock" "v3.3.2"
 github "wireapp/ono" "1.4.0"
 github "wireapp/protobuf-objc" "1.9.13"
 github "wireapp/wire-ios-cryptobox" "9.0.0"
-github "wireapp/wire-ios-data-model" "59.0.0"
+github "wireapp/wire-ios-data-model" "fix/remove-v2-calling"
 github "wireapp/wire-ios-images" "9.0.0"
 github "wireapp/wire-ios-link-preview" "5.0.1"
 github "wireapp/wire-ios-message-strategy" "40.0.0"
