--- Source/Calling/VoiceChannelV2+CallActionsInternal.swift
@@ -1,49 +1,0 @@
-//
-// Wire
-// Copyright (C) 2016 Wire Swiss GmbH
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see http://www.gnu.org/licenses/.
-//
-
-import Foundation
-import avs
-
-extension VoiceChannelV2 : CallActionsInternal {
-    
-    public var isVideoCall: Bool {
-        return conversation?.isVideoCall ?? false
-    }
-    
-    @objc(toggleVideoActive:error:)
-    public func toggleVideo(active: Bool) throws {
-        try setVideoSendActive(active)
-    }
-    
-    public func join(video: Bool) -> Bool {
-        var joined = true
-        
-        if video {
-            joined = joinVideoCall()
-        } else {
-            joined = join()
-        }
-        
-        return joined
-    }
-    
-    public func ignore() {
-        ignoreIncomingCall()
-    }
-
-}
