+++ Source/Calling/ZMCallKitDelegate.m
@@ -25,7 +25,6 @@
 #import "ZMCallKitDelegate.h"
 #import "ZMUserSession.h"
 #import "ZMUserSession+Internal.h"
-#import "VoiceChannelV2+CallFlow.h"
 #import "ZMCallKitDelegate+TypeConformance.h"
 #import <WireSyncEngine/WireSyncEngine-Swift.h>
 
@@ -310,11 +309,11 @@ NS_ASSUME_NONNULL_END
     [self.userSession performChanges:^{
         if (conversation.voiceChannel.selfUserConnectionState == VoiceChannelV2ConnectionStateNotConnected) {
             [self logInfoForConversation:conversation.remoteIdentifier.transportString line:__LINE__ format:@"CXProvider performEndCallAction: ignore incoming call"];
-            [conversation.voiceChannelRouter.currentVoiceChannel ignore];
+            [conversation.voiceChannelInternal ignore];
         }
         else {
             [self logInfoForConversation:conversation.remoteIdentifier.transportString line:__LINE__ format:@"CXProvider performEndCallAction: leave"];
-            [conversation.voiceChannelRouter.currentVoiceChannel leave];
+            [conversation.voiceChannelInternal leave];
         }
     }];
 }
@@ -407,7 +406,7 @@ NS_ASSUME_NONNULL_END
                                           update:update
                                       completion:^(NSError * _Nullable error) {
                                           if (nil != error) {
-                                              [conversation.voiceChannelRouter.currentVoiceChannel leave];
+                                              [conversation.voiceChannelInternal leave];
                                               [self logErrorForConversation:conversation.remoteIdentifier.transportString line:__LINE__ format:@"Cannot report incoming call: %@", error];
                                           } else {
                                               [self configureAudioSession];
@@ -422,7 +421,7 @@ NS_ASSUME_NONNULL_END
     
     [userSession enqueueChanges:^{
         for (ZMConversation *conversation in nonIdleCallConversations) {
-            [conversation.voiceChannelRouter.currentVoiceChannel leave];
+            [conversation.voiceChannelInternal leave];
         }
     }];
 }
@@ -512,7 +511,7 @@ NS_ASSUME_NONNULL_END
     
     [userSession performChanges:^{
         [self configureAudioSession];
-        if ([callConversation.voiceChannelRouter.currentVoiceChannel joinWithVideo:action.video]) {
+        if ([callConversation.voiceChannelInternal joinWithVideo:action.video]) {
             [action fulfill];
         } else {
             [action fail];
@@ -544,7 +543,7 @@ NS_ASSUME_NONNULL_END
     };
     
     [userSession performChanges:^{
-        if (![callConversation.voiceChannelRouter.currentVoiceChannel joinWithVideo:NO]) {
+        if (![callConversation.voiceChannelInternal joinWithVideo:NO]) {
             [action fail];
         }
     }];
