+++ Source/Calling/ZMConversation+VoiceChannel.swift
@@ -24,19 +24,19 @@ public extension ZMConversation {
     
     /// NOTE: this object is transient, and will be re-created periodically. Do not hold on to this object, hold on to the owning conversation instead.
     var voiceChannel : VoiceChannel? {
-        return voiceChannelRouter
+        return voiceChannelInternal
     }
     
     /// NOTE: this object is transient, and will be re-created periodically. Do not hold on to this object, hold on to the owning conversation instead.
-    var voiceChannelRouter : VoiceChannelRouter? {
+    var voiceChannelInternal : VoiceChannelV3? {
         
         get {
             guard conversationType == .oneOnOne || conversationType == .group else { return nil }
             
-            if let voiceChannel = objc_getAssociatedObject(self, &voiceChannelAssociatedKey) as? VoiceChannelRouter {
+            if let voiceChannel = objc_getAssociatedObject(self, &voiceChannelAssociatedKey) as? VoiceChannelV3 {
                 return voiceChannel
             } else {
-                let voiceChannel = VoiceChannelRouter(conversation: self)
+                let voiceChannel = VoiceChannelV3(conversation: self)
                 objc_setAssociatedObject(self, &voiceChannelAssociatedKey, voiceChannel, .OBJC_ASSOCIATION_RETAIN)
                 return voiceChannel
             }
