+++ Tests/Source/Synchronization/ZMSyncStrategyTests.m
@@ -701,42 +701,6 @@
     }];
 }
 
-- (void)testThatItSynchronizesCallStateChangesInUIContextToSyncContext
-{
-    // expect
-    [(id<ZMContextChangeTracker>)[self.mockUpstreamSync1 stub] objectsDidChange:OCMOCK_ANY];
-    [(id<ZMContextChangeTracker>)[self.mockUpstreamSync2 stub] objectsDidChange:OCMOCK_ANY];
-    
-    
-    // given
-    ZMConversation *uiConversation = [ZMConversation insertNewObjectInManagedObjectContext:self.uiMOC];
-    XCTAssertTrue([self.uiMOC saveOrRollback]);
-    WaitForAllGroupsToBeEmpty(0.5);
-    
-    [self.sut.syncMOC performGroupedBlockAndWaitWithReasonableTimeout:^{
-        ZMConversation *syncConversation =  (ZMConversation *)[self.sut.syncMOC objectWithID:uiConversation.objectID];
-        XCTAssertNotNil(syncConversation);
-        XCTAssertFalse(syncConversation.callDeviceIsActive);
-    }];
-
-    // when
-    uiConversation.callDeviceIsActive = YES;
-    XCTAssertTrue([self.uiMOC saveOrRollback]);
-    WaitForAllGroupsToBeEmpty(0.5);
-    // sut automagically synchronizes objects
-    
-    // then
-    XCTAssertTrue(uiConversation.callDeviceIsActive);
-
-    [self.sut.syncMOC performGroupedBlockAndWaitWithReasonableTimeout:^{
-        ZMConversation *syncConversation =  (ZMConversation *)[self.sut.syncMOC objectWithID:uiConversation.objectID];
-        XCTAssertNotNil(syncConversation);
-        XCTAssertTrue(syncConversation.callDeviceIsActive);
-    }];
-    
-    WaitForAllGroupsToBeEmpty(0.5);
-}
-
 - (void)testThatItSynchronizesChangesInSyncContextToUIContext
 {
     __block ZMUser *syncUser;
