+++ Tests/Source/UserSession/ZMUserSessionTests.m
@@ -1210,30 +1210,30 @@
     XCTAssertTrue(sender.isConnected);
 }
 
-
-- (void)testThatItCalls_DelegateShowConversationAndAcceptsCall_ZMConnectCategory
-{
-    // given
-    [self simulateLoggedInUser];
-    
-    [[[self.transportSession stub] andReturn:nil] attemptToEnqueueSyncRequestWithGenerator:OCMOCK_ANY];
-
-    UILocalNotification *note = [self notificationWithConversationForCategory:ZMIncomingCallCategory];
-    ZMConversation *conversation = [note conversationInManagedObjectContext:self.uiMOC];
-
-    [[conversation mutableOrderedSetValueForKey:@"callParticipants"] addObject:[ZMUser insertNewObjectInManagedObjectContext:self.uiMOC]];
-    [self.uiMOC saveOrRollback];
-    
-    // expect
-    [self.application setInactive];
-
-    [self checkThatItCallsTheDelegateForNotification:note responseInfo:nil actionIdentifier:nil withBlock:^(id mockDelegate) {
-        [[mockDelegate expect] showConversation:conversation];
-    }];
-    
-    // then
-    XCTAssertTrue(conversation.callDeviceIsActive);
-}
+// TODO port to calling V3
+//- (void)testThatItCalls_DelegateShowConversationAndAcceptsCall_ZMConnectCategory
+//{
+//    // given
+//    [self simulateLoggedInUser];
+//    
+//    [[[self.transportSession stub] andReturn:nil] attemptToEnqueueSyncRequestWithGenerator:OCMOCK_ANY];
+//
+//    UILocalNotification *note = [self notificationWithConversationForCategory:ZMIncomingCallCategory];
+//    ZMConversation *conversation = [note conversationInManagedObjectContext:self.uiMOC];
+//
+//    [[conversation mutableOrderedSetValueForKey:@"callParticipants"] addObject:[ZMUser insertNewObjectInManagedObjectContext:self.uiMOC]];
+//    [self.uiMOC saveOrRollback];
+//    
+//    // expect
+//    [self.application setInactive];
+//
+//    [self checkThatItCallsTheDelegateForNotification:note responseInfo:nil actionIdentifier:nil withBlock:^(id mockDelegate) {
+//        [[mockDelegate expect] showConversation:conversation];
+//    }];
+//    
+//    // then
+//    XCTAssertTrue(conversation.callDeviceIsActive);
+//}
 
 - (void)testThatIt_DoesNotAcceptsCall_ButCallsDelegateShowConversation_ZMIncomingCallCategory_NoCallParticipants
 {
